refactor: move connector consent flows from inference-api to app-api
The AgentCore Runtime data plane only proxies /invocations and /ping —
custom paths like /connectors/{id}/status returned 404 from AWS before
ever reaching the inference-api container, breaking the settings page in
cloud (worked locally only because localhost:8001 bypasses the runtime).

App-api is reachable at arbitrary HTTP paths via its ALB, so the
user-facing CRUD-ish connector endpoints live there now. The agent loop's
runtime-side OAuth flow (BeforeToolCallEvent consent hook, MCP tool
gating) is unchanged.

Code moves:
- agents/main_agent/integrations/agentcore_identity.py -> apis/shared/oauth/
- apis/inference_api/middleware/agentcore_context.py -> apis/shared/middleware/

App-api stack additions:
- AGENTCORE_RUNTIME_WORKLOAD_NAME env var (read from new SSM param)
- DYNAMODB_OAUTH_USER_TOKENS_TABLE_NAME env var
- IAM: GetWorkloadAccessToken{,ForJWT,ForUserId}, GetResourceOauth2Token,
  CompleteResourceTokenAuth (mirroring the inference-api task role so
  both APIs operate against a single shared OAuth vault).

Inference-api-stack publishes runtime-workload-identity-{arn,name} from
the auto-created workload identity (CFN attribute). Without these,
app-api can't mint workload tokens to act as the runtime workload.

Deploy ordering: Inference API now deploys before App API on fresh
environments (App API reads runtime-workload-identity-name from SSM).
Updated CLAUDE.md.

This is PR 1 of 2: connectors won't be functional in cloud until PR 2
flips the frontend call sites from inferenceApiUrl() to appApiUrl().
