+++ backend/src/agents/main_agent/session/hooks/oauth_consent.py
@@ -35,7 +35,7 @@ from strands.hooks import (
 )
 
 from agents.main_agent.integrations import oauth_token_cache
-from agents.main_agent.integrations.agentcore_identity import (
+from apis.shared.oauth.agentcore_identity import (
     CallbackUrlUnavailableError,
     WorkloadTokenUnavailableError,
     custom_parameters_for,
