--- backend/src/apis/inference_api/middleware/__init__.py
@@ -1,1 +1,0 @@
-"""Inference API middleware."""
